prov/efa: get rid of rxr_get_lower_rdm_info()
rxr_get_lower_rdm_info() convert user hints to core hints, then
pass the hints to efa_getinfo() to get a list of device rdm_info.
This device rdm_info was then converted to user rdm_info.

This patch simplified the process by directly generate
user's rdm_info from user's hint and prov_info_rxr, for
which this patch moved several function in efa_fabric.c
to efa_user_info.c and make rxr_rdm_getinfo() call them
directly

rxr_get_lower_rdm_info() and related functions were then
removed.
