+++ prov/efa/src/efa_fabric.c
@@ -64,229 +64,6 @@ const char *efa_perf_counters_str[] = {
 };
 #endif
 
-static int efa_check_hints(uint32_t version, const struct fi_info *hints,
-			   const struct fi_info *info)
-{
-	uint64_t prov_mode;
-	size_t size;
-	int ret;
-
-	if (hints->caps & ~(info->caps)) {
-		EFA_INFO(FI_LOG_CORE, "Unsupported capabilities\n");
-		OFI_INFO_CHECK(&efa_prov, info, hints, caps, FI_TYPE_CAPS);
-		return -FI_ENODATA;
-	}
-
-	prov_mode = ofi_mr_get_prov_mode(version, hints, info);
-
-	if ((hints->mode & prov_mode) != prov_mode) {
-		EFA_INFO(FI_LOG_CORE, "Required hints mode bits not set\n");
-		OFI_INFO_MODE(&efa_prov, prov_mode, hints->mode);
-		return -FI_ENODATA;
-	}
-
-	if (hints->fabric_attr) {
-		ret = ofi_check_fabric_attr(&efa_prov, info->fabric_attr,
-					    hints->fabric_attr);
-
-		if (ret)
-			return ret;
-	}
-
-	switch (hints->addr_format) {
-	case FI_FORMAT_UNSPEC:
-	case FI_ADDR_EFA:
-		size = EFA_EP_ADDR_LEN;
-		break;
-	default:
-		EFA_INFO(FI_LOG_CORE,
-			 "Address format not supported: hints[%u], supported[%u,%u]\n",
-			 hints->addr_format, FI_FORMAT_UNSPEC, FI_ADDR_EFA);
-		return -FI_ENODATA;
-	}
-
-	if (hints->src_addr && hints->src_addrlen < size)
-		return -FI_ENODATA;
-
-	if (hints->dest_addr && hints->dest_addrlen < size)
-		return -FI_ENODATA;
-
-	if (hints->domain_attr) {
-		ret = ofi_check_domain_attr(&efa_prov, version, info->domain_attr, hints);
-		if (ret)
-			return ret;
-	}
-
-	if (hints->ep_attr) {
-		ret = ofi_check_ep_attr(&efa_util_prov, info->fabric_attr->api_version, info, hints);
-		if (ret)
-			return ret;
-	}
-
-	if (hints->rx_attr) {
-		ret = ofi_check_rx_attr(&efa_prov, info, hints->rx_attr, hints->mode);
-		if (ret)
-			return ret;
-	}
-
-	if (hints->tx_attr) {
-		ret = ofi_check_tx_attr(&efa_prov, info->tx_attr, hints->tx_attr, hints->mode);
-		if (ret)
-			return ret;
-	}
-
-	return 0;
-}
-
-static int efa_str_to_ep_addr(const char *node, const char *service, struct efa_ep_addr *addr)
-{
-	int ret;
-
-	if (!node)
-		return -FI_EINVAL;
-
-	memset(addr, 0, sizeof(*addr));
-
-	ret = inet_pton(AF_INET6, node, addr->raw);
-	if (ret != 1)
-		return -FI_EINVAL;
-	if (service)
-		addr->qpn = atoi(service);
-
-	return 0;
-}
-
-static int efa_node_matches_addr(struct efa_ep_addr *addr, const char *node)
-{
-	struct efa_ep_addr eaddr;
-
-	efa_str_to_ep_addr(node, NULL, &eaddr);
-	return memcmp(&eaddr.raw, &addr->raw, sizeof(addr->raw));
-}
-
-static int efa_get_matching_info(uint32_t version, const char *node, uint64_t flags,
-				 const struct fi_info *hints, struct fi_info **info)
-{
-	const struct fi_info *check_info;
-	struct fi_info *fi, *tail;
-	int ret;
-
-	*info = tail = NULL;
-
-	for (check_info = efa_util_prov.info; check_info; check_info = check_info->next) {
-		ret = 0;
-		if (flags & FI_SOURCE) {
-			if (node)
-				ret = efa_node_matches_addr(check_info->src_addr, node);
-		} else if (hints && hints->src_addr) {
-			ret = memcmp(check_info->src_addr, hints->src_addr, EFA_EP_ADDR_LEN);
-		}
-
-		if (ret)
-			continue;
-		EFA_INFO(FI_LOG_FABRIC, "found match for interface %s %s\n", node, check_info->fabric_attr->name);
-		if (hints) {
-			ret = efa_check_hints(version, hints, check_info);
-			if (ret)
-				continue;
-		}
-
-		fi = fi_dupinfo(check_info);
-		if (!fi) {
-			ret = -FI_ENOMEM;
-			goto err_free_info;
-		}
-
-		fi->fabric_attr->api_version = version;
-
-		if (!*info)
-			*info = fi;
-		else
-			tail->next = fi;
-		tail = fi;
-	}
-
-	if (!*info)
-		return -FI_ENODATA;
-
-	return 0;
-
-err_free_info:
-	fi_freeinfo(*info);
-	*info = NULL;
-	return ret;
-}
-
-static int efa_set_fi_address(const char *node, const char *service, uint64_t flags,
-			      const struct fi_info *hints, struct fi_info *fi)
-{
-	struct efa_ep_addr tmp_addr;
-	void *dest_addr = NULL;
-	int ret = FI_SUCCESS;
-	struct fi_info *cur;
-
-	if (flags & FI_SOURCE) {
-		if (hints && hints->dest_addr)
-			dest_addr = hints->dest_addr;
-	} else {
-		if (node || service) {
-			ret = efa_str_to_ep_addr(node, service, &tmp_addr);
-			if (ret)
-				return ret;
-			dest_addr = &tmp_addr;
-		} else if (hints && hints->dest_addr) {
-			dest_addr = hints->dest_addr;
-		}
-	}
-
-	if (dest_addr) {
-		for (cur = fi; cur; cur = cur->next) {
-			cur->dest_addr = malloc(EFA_EP_ADDR_LEN);
-			if (!cur->dest_addr) {
-				for (; fi->dest_addr; fi = fi->next)
-					free(fi->dest_addr);
-				return -FI_ENOMEM;
-			}
-			memcpy(cur->dest_addr, dest_addr, EFA_EP_ADDR_LEN);
-			cur->dest_addrlen = EFA_EP_ADDR_LEN;
-		}
-	}
-	return ret;
-}
-
-int efa_getinfo(uint32_t version, const char *node, const char *service,
-		uint64_t flags, const struct fi_info *hints, struct fi_info **info)
-{
-	int ret;
-
-	if (!(flags & FI_SOURCE) && hints && hints->src_addr &&
-	    hints->src_addrlen != EFA_EP_ADDR_LEN)
-		return -FI_ENODATA;
-
-	if (((!node && !service) || (flags & FI_SOURCE)) &&
-	    hints && hints->dest_addr &&
-	    hints->dest_addrlen != EFA_EP_ADDR_LEN)
-		return -FI_ENODATA;
-
-	ret = efa_get_matching_info(version, node, flags, hints, info);
-	if (ret)
-		goto out;
-
-	ret = efa_set_fi_address(node, service, flags, hints, *info);
-	if (ret)
-		goto out;
-
-	ofi_alter_info(*info, hints, version);
-
-out:
-	if (!ret || ret == -FI_ENOMEM || ret == -FI_ENODEV) {
-		return ret;
-	} else {
-		fi_freeinfo(*info);
-		*info = NULL;
-		return -FI_ENODATA;
-	}
-}
 
 static int efa_fabric_close(fid_t fid)
 {
