+++ prov/efa/src/efa_prov_info.h
@@ -42,5 +42,7 @@ int efa_prov_info_alloc(struct fi_info **prov_info,
 int efa_prov_info_alloc_for_rxr(struct fi_info **prov_info_rxr,
 				struct efa_device *device);
 
+int efa_prov_info_compare_src_addr(const char *node, uint64_t flags, const struct fi_info *hints,
+				   const struct fi_info *fi);
 
 #endif
